cms-alfresco-platform-jar/src/main/java/gov/uspto/trademark/cms/repo/webscripts/healthstatus/Detail.java
package gov.uspto.trademark.cms.repo.webscripts.healthstatus;

import org.codehaus.jackson.annotate.JsonProperty;
import org.codehaus.jackson.annotate.JsonPropertyOrder;
import org.codehaus.jackson.map.annotate.JsonSerialize;

@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
@JsonPropertyOrder({ "name", "status", "details" })
public class Detail {

    @JsonProperty("name")
    private String name;
    @JsonProperty("status")
    private String status;
    @JsonProperty("details")
    private String details;

    /**
     *
     * @return The name
     */
    @JsonProperty("name")
    public String getName() {
        return name;
    }

    /**
     *
     * @param name
     *            The name
     */
    @JsonProperty("name")
    public void setName(String name) {
        this.name = name;
    }

    /**
     *
     * @return The status
     */
    @JsonProperty("status")
    public String getStatus() {
        return status;
    }

    /**
     *
     * @param status
     *            The status
     */
    @JsonProperty("status")
    public void setStatus(String status) {
        this.status = status;
    }

    /**
     *
     * @return The details
     */
    @JsonProperty("details")
    public String getDetails() {
        return details;
    }

    /**
     *
     * @param details
     *            The details
     */
    @JsonProperty("details")
    public void setDetails(String details) {
        this.details = details;
    }

}